bind FullSnapshotReport service
